refactor(controller): change get value in controller with argument resolver & static method
